Consolodate files
