--- thread/timer.h
@@ -1,27 +1,0 @@
-#ifndef TIMER_H
-#define TIMER_H
-
-#include <stdint.h>
-
-class TimerInterrupt; // forward declaration
-class pruThread; // forward declaration
-
-class pruTimer
-{
-friend class TimerInterrupt;
-
-private:
-
-    TimerInterrupt*     interruptPtr;
-    uint8_t             slice;
-    pruThread*          timerOwnerPtr;
-
-    void startTimer(void);
-    void timerTick();           // Private timer tiggered method
-
-public:
-    pruTimer(uint8_t slice, pruThread* ownerPtr);
-
-};
-
-#endif
